+++ FileOperation/FileOperationEN.java
@@ -93,6 +93,8 @@ public class FileOperationEN {
 
         // Different file extensions 
         chooser = new JFileChooser();
+        chooser.addChoosableFileFilter(new FileFilterEN("*", "All Files"));
+        chooser.addChoosableFileFilter(new FileFilterEN(".txt", "Text Files(*.txt)"));
         chooser.addChoosableFileFilter(new FileFilterEN(".java", "Java Source Files(*.java)"));
         chooser.addChoosableFileFilter(new FileFilterEN(".py", "Python Files(*.py)"));
         chooser.addChoosableFileFilter(new FileFilterEN(".c", "C Programming Language(.c)"));
@@ -101,12 +103,10 @@ public class FileOperationEN {
 		chooser.addChoosableFileFilter(new FileFilterEN(".d", "D Programming Language(.d)"));
 		chooser.addChoosableFileFilter(new FileFilterEN(".sh", "Shell Script File(*.sh)"));
 		chooser.addChoosableFileFilter(new FileFilterEN(".bat", "Batch File(*.bat)"));
-        chooser.addChoosableFileFilter(new FileFilterEN(".txt", "Text Files(*.txt)"));
 		chooser.addChoosableFileFilter(new FileFilterEN(".rtf", "Rich Text Format(*.rtf)"));
         chooser.addChoosableFileFilter(new FileFilterEN(".pdf", "Portable Document Files(*.pdf)"));
         chooser.addChoosableFileFilter(new FileFilterEN(".html", "Hyper Text Markup Language(*.html)"));
         chooser.addChoosableFileFilter(new FileFilterEN(".asm", "Assembler(*.asm)"));
-		chooser.addChoosableFileFilter(new FileFilterEN("*", "All Files"));
         chooser.setCurrentDirectory(new File("."));
     }
 //////////////////////////////////////
@@ -263,9 +263,6 @@ public class FileOperationEN {
     ///////////////////////////////////////
     public void newFile() {
         if (!confirmSave()) return;
-        
-        new FNotepadEN(true);
-
         this.npd.ta.setText("");
         fileName = new String("Untitled");
         fileRef = new File(fileName);